Changed the handling of preconditioners in `inverse` LinearOperators.
The option `pc` must now take as value a `LinearOperator`, which is the preconditioner.
If `pc=None`, an `IdentityOperator` of the domain space is created automatically.
The class `InverseLinearOperators` could be deprecated, since the Jacobi preconditiner should be implemented as a `LinearOperator`.

The new solver `PBiConjugateGradientStabilized` (preconditioned) has been added from Struphy.

`test_solvers.py` has been updated. The Jacobi preconditioner is not tested at the moment.